refactor(page-forum): refactor components into separate files and add toggle button animation
