+++ src/pages/Forum.tsx
@@ -1,7 +1,11 @@
-import { useState } from "react";
-import { Button } from "@/components/ui/button";
-import { ImagePlus, Search, SlidersHorizontal } from "lucide-react";
-import { TagPopuler, Card, Sidebar, ResponsiveCard, ResponsivePost } from "@/components/ui/forum"; // pastikan path ini benar
+import { ResponsivePost } from "@/components/modules/Forum/ResponsivePost";
+import { ResponsiveCard } from "@/components/modules/Forum/ResponsiveCard";
+import { Sidebar } from "@/components/modules/Forum/Sidebar";
+import { LeftBar } from "@/components/modules/Forum/LeftBar";
+import { RightBar } from "@/components/modules/Forum/RightBar";
+
+import { cn } from "@/lib/utils";
+import { useScrollPosition } from "@/hooks/use-scroll-position";
 
 interface TagData {
   hashtag: string;
@@ -29,6 +33,8 @@ interface Comment {
 }
 
 export default function Forum() {
+  const isScrolled = useScrollPosition();
+
   const tagData: TagData[] = [
     { hashtag: "#haircare", posted: "1.067 diposting" },
     { hashtag: "#hair", posted: "837 diposting" },
@@ -104,13 +110,18 @@ export default function Forum() {
   ];
 
   return (
-    <div className="container flex flex flex-col justify-center gap-2 bg-shark-900 p-0 lg:container lg:bg-white lg:pb-[30px] lg:pt-8">
-      <div className="fixed left-0 top-24 block flex h-screen w-12 justify-center border-r-[0.1875rem] border-t-[0.1875rem] border-shark-800 bg-shark-900 lg:hidden">
+    <div className=" flex flex-col justify-center gap-2 bg-shark-900 p-0 xl:container lg:pt-8 xl:bg-white xl:pb-[30px]">
+      <div
+        className={cn(
+          "fixed left-0 flex h-screen w-12 justify-center border-r-[0.1875rem] border-t-[0.1875rem] border-shark-800 bg-shark-900 xl:hidden",
+          isScrolled ? "top-0" : "top-24"
+        )}
+      >
         <div>
           <Sidebar />
         </div>
       </div>
-      <div className="ml-12 block h-full w-full lg:hidden">
+      <div className="ml-0 block h-full w-full pl-6 md:px-12 xl:ml-12 xl:hidden">
         <ResponsivePost />
         {posts.map((post: Post) => (
           <ResponsiveCard
@@ -120,7 +131,7 @@ export default function Forum() {
           />
         ))}
       </div>
-      <div className="hidden lg:block">
+      <div className="hidden xl:block">
         <LeftBar tagData={tagData} />
         <div className="ml-36">
           <RightBar posts={posts} comments={comments} />
@@ -129,102 +140,3 @@ export default function Forum() {
     </div>
   );
 }
-
-const LeftBar = ({ tagData }: { tagData: TagData[] }) => (
-  <div className="w-62 static fixed left-24 top-[8rem] space-y-3.5">
-    <div className="relative w-full rounded-[0.625rem] bg-shark-900 p-5">
-      <input
-        type="text"
-        placeholder="Cari Topik..."
-        className="w-full rounded-[0.625rem] border border-gray-300 py-1.5 pl-3 pr-10 focus:outline-none"
-      />
-      <button className="absolute inset-y-0 right-4 flex items-center pr-3">
-        <Search color="#c7c7c7" />
-      </button>
-    </div>
-    <div className="relative w-full rounded-[0.625rem] bg-shark-900 p-2.5">
-      <div className="flex flex-col gap-1 p-2.5">
-        <div className="flex w-fit gap-2 border-b-2 border-laser-300">
-          <SlidersHorizontal color="#cdb16e" />
-          <h1 className="text-xl font-semibold text-laser-300">Filter</h1>
-        </div>
-        <div className="flex flex-col gap-1 text-lg font-extrabold text-white">
-          <a href="">
-            <h3>Terbaru dan Terkini</h3>
-          </a>
-          <a href="">
-            <h3>Sedang Tren</h3>
-          </a>
-          <a href="">
-            <h3>Terlama</h3>
-          </a>
-        </div>
-      </div>
-    </div>
-    <div>
-      <TagPopuler
-        hashtags={tagData.map((data: TagData) => data.hashtag)}
-        posteds={tagData.map((data: TagData) => data.posted)}
-      />
-    </div>
-  </div>
-);
-
-const RightBar = ({ posts, comments }: { posts: Post[]; comments: Comment[] }) => {
-  const [visibleComments, setVisibleComments] = useState<{
-    [key: number]: boolean;
-  }>({});
-
-  const toggleComments = (postId: number) => {
-    setVisibleComments(prev => ({ ...prev, [postId]: !prev[postId] }));
-  };
-
-  return (
-    <div className="">
-      <div className="inline-flex w-full justify-center">
-        <div className="item-center flex h-[6.875rem] w-[47.5rem] justify-center gap-2.5 rounded-[0.625rem] bg-[#2E323A] pb-10 pt-5">
-          <figure>
-            <img src="/images/Ellipse-2.png" alt="" />
-          </figure>
-          <div className="space-y-4">
-            <div>
-              <input
-                type="text"
-                placeholder="Let's share what you think about hairstyles ..."
-                className="h-[0.625rem] w-[36.25rem] rounded-[0.625rem] p-[1.1rem] focus:outline-none"
-              />
-            </div>
-            <div>
-              <a href="">
-                <div className="flex gap-[0.38rem]">
-                  <ImagePlus color="#cdb16e" />
-                  <h1 className="font-medium text-white">Photo</h1>
-                </div>
-              </a>
-            </div>
-          </div>
-          <div>
-            <Button variant="post" size="post">
-              Post
-            </Button>
-          </div>
-        </div>
-      </div>
-      {posts.map(post => (
-        <Card
-          key={post.id}
-          profileImage={post.profileImage}
-          name={post.name}
-          text={post.text}
-          image={post.image || ""}
-          hashtags={post.hashtags}
-          comments={post.comments}
-          likes={post.likes}
-          showComments={!!visibleComments[post.id]}
-          toggleComments={() => toggleComments(post.id)}
-          commentData={comments.filter(comment => comment.postId === post.id)}
-        />
-      ))}
-    </div>
-  );
-};
